# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/waitlist": {
      "post": {
        "operationId": "joinWaitlist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
    "/api/tenants": {
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "x-bootstrap-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
    "/api/tenants/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
    "/api/tenants/{id}/entities": {
      "get": {
        "operationId": "entities",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
    "/api/entities": {
      "get": {
        "operationId": "listEntities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      },
      "post": {
        "operationId": "createEntity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
...
    "/api/entities/{id}": {
      "patch": {
        "operationId": "updateEntity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
...
    "/api/plan": {
      "get": {
        "operationId": "getPlan",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      },
      "patch": {
        "operationId": "updatePlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
...
    "/api/modules/{key}": {
      "patch": {
        "operationId": "setModule",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
    "/api/ai-settings": {
      "get": {
        "operationId": "getAiSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      },
      "patch": {
        "operationId": "updateAiSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAiSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tenancy"
        ]
      }
    },
...
    "/api/auth/workspace": {
      "get": {
        "operationId": "workspace",
        "parameters": [
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resolve a workspace by slug or tenant id",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Procure Ledger API",
    "description": "Phase 1 modular monolith — auth session, master data, invoices, capture/files, integration jobs. Authenticate with a session bearer token or a `pl_…` API key (legacy `aptora_…` keys still accepted).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT / session / pl_ API key",
        "type": "http",
        "description": "Session token from POST /api/auth/login (or invite accept), or an API key starting with pl_ (legacy aptora_ accepted)"
      }
    },
    "schemas": {
      "WaitlistDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9-]+$"
          },
          "region": {
            "type": "string",
            "enum": [
              "us",
              "eu"
            ]
          },
          "setupToken": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "CreateEntityDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEntityDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateModuleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAiSettingsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}